refactor: 회차별 정보 크롤링 리팩토링
